requirements.txt
# Core
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.38.0
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart>=0.0.9

# SQL & dbt
sqlglot==25.24.0
dbt-core==1.8.7
dbt-duckdb==1.8.4

# Lineage graph
networkx==3.3
matplotlib==3.9.2

# LLM & Agents
anthropic>=0.84.0
langgraph>=1.1.0
langchain-anthropic>=1.4.0

# HTTP client (for dbt Cloud API)
httpx>=0.27.0

# Production ops
slowapi>=0.1.9
structlog>=25.0.0

# Persistence (SQLite for dev, Postgres-ready for prod)
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.22.0
greenlet>=3.0.0

# Utilities
pandas==2.2.3
pytest==9.0.2
